src/numeraifold/domains/identification.py
"""
Module for identifying feature domains and converting tabular data into sequence-like representations.

This module includes:
- FeatureDomainIdentifier: A placeholder class for future extensions in feature domain identification.
- identify_feature_domains: Clusters features into natural domains using PCA, UMAP, and KMeans.
- create_sequence_representation: Converts a DataFrame into a sequence representation with features organized by domain.
"""

# Import necessary libraries
import numpy as np
import pandas as pd
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA
from sklearn.cluster import KMeans
import umap
from tqdm import tqdm

class FeatureDomainIdentifier:
    """
    Placeholder class for feature domain identification.

    This class is reserved for potential future methods or enhancements
    in the identification of feature domains.
    """
    pass

def identify_feature_domains(df, features, n_clusters=10, random_state=42):
    """
    Identify natural feature domains using dimensionality reduction and clustering.

    This function standardizes the input features, reduces their dimensionality
    with PCA (and UMAP as a fallback), and clusters them into domains using KMeans.
    It includes error handling to return safe fallback values in case of issues.

    Parameters:
        df (pd.DataFrame): DataFrame containing the input data.
        features (list): List of feature column names from the DataFrame.
        n_clusters (int): Desired number of clusters/domains (default: 10).
        random_state (int): Seed for reproducibility (default: 42).

    Returns:
        tuple: A tuple containing:
            - feature_groups (dict): Dictionary mapping domain names to lists of features.
            - embedding (np.ndarray): 2D embedding of features after dimensionality reduction.
            - cluster_labels (np.ndarray): Cluster labels for each feature.
            - pca_result (np.ndarray): PCA result before UMAP reduction.
    """
    print("Identifying feature domains...")

    try:
        # Input validation: Ensure at least two features are provided for clustering.
        if len(features) < 2:
            print(f"Warning: Only {len(features)} features available. Need at least 2 for clustering.")
            return {"domain_0": features}, None, np.zeros(len(features)), None

        # Initialize variables for embedding and cluster labels.
        embedding = None
        cluster_labels = None

        # Standardize features and handle missing values by replacing them with zeros.
        feature_matrix = df[features].fillna(0).values
        scaler = StandardScaler()
        feature_matrix = scaler.fit_transform(feature_matrix)

        # Perform PCA for initial dimensionality reduction.
        n_components = min(20, len(features) - 1)
        pca = PCA(n_components=n_components, random_state=random_state)
        pca_result = pca.fit_transform(feature_matrix)

        # Further reduce dimensions using UMAP.
        try:
            reducer = umap.UMAP(
                n_neighbors=min(15, len(features) - 1),
                min_dist=0.1,
                n_components=2,
                random_state=random_state
            )
            embedding = reducer.fit_transform(pca_result)
        except Exception as e:
            print(f"UMAP reduction failed: {e}. Using first 2 PCA components.")
            # Fall back to the first 2 PCA components if UMAP fails.
            embedding = pca_result[:, :2]

        # Adjust number of clusters if necessary.
        n_clusters = min(n_clusters, len(features))
        kmeans = KMeans(n_clusters=n_clusters, random_state=random_state)
        cluster_labels = kmeans.fit_predict(embedding)

        # Group features by their cluster labels to form domains.
        feature_groups = {}
        for cluster_id in range(n_clusters):
            indices = np.where(cluster_labels == cluster_id)[0]
            domain_name = f"domain_{cluster_id}"
            feature_groups[domain_name] = [features[i] for i in indices]

        print(f"Successfully identified {len(feature_groups)} feature domains")
        return feature_groups, embedding, cluster_labels, pca_result

    except Exception as e:
        print(f"Error in domain identification: {e}")
        # Return safe fallback values in case of an error.
        feature_groups = {"domain_0": features}
        cluster_labels = np.zeros(len(features))
        embedding = np.random.randn(len(features), 2)  # Simple 2D random embedding.
        return feature_groups, embedding, cluster_labels, None

def create_sequence_representation(df, feature_groups, era_col='era'):
    """
    Transform tabular data into a sequence-like representation.

    Each row in the DataFrame becomes a sequence with features organized into domains.
    If the specified era column is missing, artificial eras are generated based on index chunks.
    The function processes data in batches (by era) for improved memory efficiency and includes
    error handling to deal with any issues during feature extraction.

    Parameters:
        df (pd.DataFrame): DataFrame containing the input data.
        feature_groups (dict): Dictionary mapping domain names to lists of features.
        era_col (str): Column name representing the era or group in the data (default: 'era').

    Returns:
        dict: Dictionary where each key is a tuple (era, row index) and each value is a
              dictionary mapping domain names to lists of feature values.
    """
    print("Creating sequence representation...")

    # Validate if the era column exists; if not, create artificial eras.
    if era_col not in df.columns:
        print(f"Warning: '{era_col}' not found in DataFrame. Using index groups.")
        chunk_size = 1000
        df = df.copy()
        df[era_col] = df.index // chunk_size

    # If no feature groups are provided, create a single group with all columns except the era.
    if not feature_groups:
        print("Warning: Empty feature groups provided. Creating single domain with all features.")
        all_features = df.columns.drop(era_col).tolist()
        feature_groups = {"domain_all": all_features}

    # Get unique eras from the DataFrame.
    eras = df[era_col].unique()

    # Initialize a dictionary to hold the sequence representations.
    sequences = {}

    # Process data for each era for memory efficiency.
    for era in tqdm(eras, desc="Processing eras"):
        era_df = df[df[era_col] == era]

        for idx, row in era_df.iterrows():
            # Initialize the sequence for the current row.
            sequence = {}
            for domain, domain_features in feature_groups.items():
                # Filter out any features that are not present in the DataFrame.
                valid_features = [f for f in domain_features if f in df.columns]
                if not valid_features:
                    continue

                # Retrieve feature values, replacing missing values with zeros.
                try:
                    domain_values = row[valid_features].fillna(0).tolist()
                    sequence[domain] = domain_values
                except Exception as e:
                    print(f"Error processing row {idx}, domain {domain}: {e}")
                    # Fallback: use zeros if an error occurs.
                    sequence[domain] = [0.0] * len(valid_features)

            # Only store sequences that are not empty, with a compound key of (era, index).
            if sequence:
                sequences[(era, idx)] = sequence

    if not sequences:
        print("Warning: No valid sequences created. Check feature groups and data.")

    return sequences
