requirements.txt
torch>=1.9.0
numpy>=1.19.0
pandas>=1.3.0
scikit-learn>=0.24.0
umap-learn>=0.5.0
hdbscan>=0.8.0
lightgbm>=3.2.0
tqdm>=4.62.0
matplotlib>=3.4.0
seaborn>=0.11.0
numerapi>=2.8.0
numerblox>=0.1.0  # For Numerai competition specific features
plotly>=5.3.0  # Optional for interactive visualizations